pyproject.toml
[project]
name = "sqlmulti"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "adbc-driver-postgresql>=1.4.0",
    "polars>=1.22.0",
    "psycopg2-binary>=2.9.10",
    "pyarrow>=19.0.0",
    "ruff>=0.6.9",
    "sqlalchemy>=2.0.38",
]
